Improve CLI usage
This commit brings a new main class to HdrLogProcessing.  It acts like a
command dispatcher so that the users do not need to remember the class
name they want to invoke.  The uberjar is also made executable by
specifying that dispatcher as the `Main-Class` in `MANIFEST.MF`.

The old `java -cp` way still works, regardless.